+++ HISTORY.rst
@@ -1,6 +1,13 @@
 History/Changelog
 -----------------
 
+0.8.1: 2014-01-26
+~~~~~~~~~~~~~~~~~
+
+- Add documentation for using Two Factor Authentication
+
+- Fix oversight where ``github3.login`` could not be used for 2FA
+
 0.8.0: 2014-01-03
 ~~~~~~~~~~~~~~~~~